feat: 增加公司管理字段

+++ server/src/main/java/com/avenger/jt808/domain/repository/CircularAreaRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
  * Created by jg.wang on 2020/5/9.
  * Description:
  */
-public interface CircularAreaRepository extends JpaRepository<CircularArea, Long>, JpaSpecificationExecutor<CircularArea> {
+public interface CircularAreaRepository extends JpaRepository<CircularArea, Integer>, JpaSpecificationExecutor<CircularArea> {
 }
